# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/accesses": {
      "get": {
        "operationId": "AccessesController_list",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar acessos faciais (MongoDB), por empresa",
        "tags": [
          "accesses"
        ]
      }
    },
...
    "/member/me": {
      "get": {
        "operationId": "MemberPortalController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Perfil do membro autenticado (app)",
        "tags": [
          "member-portal"
        ]
      }
    },
    "/member/accesses": {
      "get": {
        "operationId": "MemberPortalController_listAccesses",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Histórico de acessos faciais do membro autenticado",
        "tags": [
          "member-portal"
        ]
      }
    },
    "/member/accesses/snapshot": {
      "get": {
        "operationId": "MemberPortalController_accessSnapshot",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Proxy da foto de captura do evento (URL deve ser de leitor Intelbras da escola)",
        "tags": [
          "member-portal"
        ]
      }
    },
...
    "/responsible/managed-responsibles": {
      "post": {
        "operationId": "ManagedResponsiblesController_createManaged",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManagedResponsibleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cadastrar responsável presencialmente (Fluxo 1)",
        "tags": [
          "managed-responsibles"
        ]
      },
      "get": {
        "operationId": "ManagedResponsiblesController_listManaged",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar responsáveis do núcleo familiar",
        "tags": [
          "managed-responsibles"
        ]
      }
    },
    "/responsible/managed-responsibles/{id}": {
      "delete": {
        "operationId": "ManagedResponsiblesController_deleteManaged",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Excluir responsável do núcleo familiar (apenas pai/mãe)",
        "tags": [
          "managed-responsibles"
        ]
      }
    },
    "/responsible/responsible-invitations": {
      "post": {
        "operationId": "ManagedResponsiblesController_createInvitation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManagedResponsibleInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar convite de cadastro de responsável (Fluxo 2)",
        "tags": [
          "managed-responsibles"
        ]
      },
      "get": {
        "operationId": "ManagedResponsiblesController_listInvitations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar convites de cadastro de responsável",
        "tags": [
          "managed-responsibles"
        ]
      }
    },
    "/responsible/responsible-invitations/{id}/face-url": {
      "get": {
        "operationId": "ManagedResponsiblesController_faceUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "URL assinada da foto enviada pelo convidado",
        "tags": [
          "managed-responsibles"
        ]
      }
    },
    "/responsible/responsible-invitations/{id}/approve-face": {
      "post": {
        "operationId": "ManagedResponsiblesController_approveFace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aprovar face do convidado",
        "tags": [
          "managed-responsibles"
        ]
      }
    },
    "/responsible/responsible-invitations/{id}/reject-face": {
      "post": {
        "operationId": "ManagedResponsiblesController_rejectFace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Recusar face do convidado",
        "tags": [
          "managed-responsibles"
        ]
      }
    },
    "/responsible/responsible-invitations/{id}/approve-plate": {
      "post": {
        "operationId": "ManagedResponsiblesController_approvePlate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aprovar placa do convidado",
        "tags": [
          "managed-responsibles"
        ]
      }
    },
    "/responsible/responsible-invitations/{id}/reject-plate": {
      "post": {
        "operationId": "ManagedResponsiblesController_rejectPlate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Recusar placa do convidado",
        "tags": [
          "managed-responsibles"
        ]
      }
    },
    "/responsible/responsible-invitations/{id}/cancel": {
      "patch": {
        "operationId": "ManagedResponsiblesController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancelar convite pendente",
        "tags": [
          "managed-responsibles"
        ]
      }
    },
...
    "/me/navigation": {
      "get": {
        "operationId": "MeController_navigation",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Paths permitidos na sidebar da área empresa",
        "tags": [
          "me"
        ]
      }
    },
    "/me/can-check": {
      "get": {
        "operationId": "MeController_canCheck",
        "parameters": [
          {
            "name": "feature",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Checa permissão granular (empresa). Usado pelo middleware Next.js",
        "tags": [
          "me"
        ]
      }
    },
    "/me/permissions/{featureSlug}": {
      "get": {
        "operationId": "MeController_permissionsForFeature",
        "parameters": [
          {
            "name": "featureSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ações efetivas para uma feature (equivalente a getPermissions no Next.js)",
        "tags": [
          "me"
        ]
      }
    },
...
  },
  "info": {
    "title": "Face2go API",
    "description": "Backend NestJS da plataforma Face2go",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateManagedResponsibleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "phone": {
            "type": "string",
            "maxLength": 32,
            "nullable": true
          },
          "document": {
            "type": "string",
            "maxLength": 32,
            "nullable": true
          },
          "isActive": {
            "default": true,
            "type": "boolean"
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "nullable": true
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "nullable": true
          },
          "linkedResponsibleId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "students": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "studentId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "relationshipType": {
                  "type": "string",
                  "enum": [
                    "parent",
                    "grandparent",
                    "aunt_uncle",
                    "sibling",
                    "godparent",
                    "guardian",
                    "other"
                  ]
                },
                "isAuthorizedPickup": {
                  "default": true,
                  "type": "boolean"
                }
              },
              "required": [
                "studentId",
                "relationshipType"
              ]
            }
          },
          "imageBase64": {
            "type": "string",
            "minLength": 64
          },
          "vehicle": {
            "type": "object",
            "properties": {
              "plate": {
                "type": "string"
              },
              "brand": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "model": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "color": {
                "type": "string",
                "minLength": 1,
                "maxLength": 50
              }
            },
            "required": [
              "plate",
              "brand",
              "model",
              "color"
            ]
          }
        },
        "required": [
          "name",
          "students"
        ]
      },
      "CreateManagedResponsibleInvitationDto": {
        "type": "object",
        "properties": {
          "students": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "studentId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "relationshipType": {
                  "type": "string",
                  "enum": [
                    "parent",
                    "grandparent",
                    "aunt_uncle",
                    "sibling",
                    "godparent",
                    "guardian",
                    "other"
                  ]
                },
                "isAuthorizedPickup": {
                  "default": true,
                  "type": "boolean"
                }
              },
              "required": [
                "studentId",
                "relationshipType"
              ]
            }
          }
        },
        "required": [
          "students"
        ]
      },
...
    }
  }
}